+++ Src/SOS.lib/Models/Processed/DataStewardship/Event/WeatherMeasuring.cs
@@ -15,7 +15,7 @@ namespace SOS.Lib.Models.Processed.DataStewardship.Event
         /// </summary>
         [Required]
         [DataMember(Name = "weatherMeasure")]
-        public decimal? WeatherMeasure { get; set; }
+        public double? WeatherMeasure { get; set; }
 
         /// <summary>
         /// Unit for a reported measurement (given in the attribute "vädermått").
